better testing for these cases
