+++ crates/biome_service/src/file_handlers/html/parse_embedded_nodes.rs
@@ -1036,3 +1036,7 @@ fn register_svelte_destructured_bindings(
 
     Some(())
 }
+
+#[cfg(test)]
+#[path = "parse_embedded_nodes.tests.rs"]
+mod tests;
